exercise 4.22

+++ part4/blog/controllers/blogs.js
@@ -39,6 +39,15 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), 'secret');
+  // console.log('here');
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' });
+  // }
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() != decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'not owner' });
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
